slimbus: Allow controller to override default scheduling algorithm
Framework provides a slimbus scheduling algorithm and controller may
choose to override it. e.g. controller will override this if it
supports fixed set of channels at a fixed clock gear.
In that case, it is controller's responsibility to act on client
requests and update channel scheduling, subframe coding information.

Change-Id: I31fe72ef85f931900467b737652da7516d173dc5
Signed-off-by: Sagar Dharia <sdharia@codeaurora.org>

Conflicts:
	include/linux/slimbus/slimbus.h

Conflicts:
	include/linux/slimbus/slimbus.h